With extractor validation
