+++ src/main/java/io/github/jdbctemplatemapper/core/IRelationshipBuilder.java
@@ -1,4 +1,4 @@
-package io.github.jdbctemplatemapper.support;
+package io.github.jdbctemplatemapper.core;
 
 public interface IRelationshipBuilder<T> {
     IMapperExtractionBuilder<T>  hasMany(Class<?> relatedClazz, String propertyName);
