+++ src/main/java/io/github/jdbctemplatemapper/core/RelationshipType.java
@@ -1,0 +1,6 @@
+package io.github.jdbctemplatemapper.core;
+
+enum RelationshipType {
+    HAS_ONE,
+    HAS_MANY; 
+}
