+++ src/main/java/io/github/jdbctemplatemapper/exception/MapperExtractorException.java
@@ -1,0 +1,9 @@
+package io.github.jdbctemplatemapper.exception;
+
+public class MapperExtractorException extends RuntimeException{
+    private static final long serialVersionUID = 1L;
+
+    public MapperExtractorException(String message) {
+        super(message);
+    }
+}
