--- src/main/java/io/github/jdbctemplatemapper/support/RelationshipType.java
@@ -1,6 +1,0 @@
-package io.github.jdbctemplatemapper.support;
-
-public enum RelationshipType {
-    HAS_ONE,
-    HAS_MANY; 
-}
